Extract Code and improve naming for a missing peg
