+++ CodeBreaker/CodeBreaker/CodeBreaker.swift
@@ -9,6 +9,10 @@ import Foundation
 
 typealias Peg = String
 
+extension Peg {
+    static let missing = ""
+}
+
 struct Theme {
     var name: String
     var pegs: [Peg]
@@ -62,64 +66,7 @@ struct CodeBreaker {
             let newPeg = pegChoices[(indexOfExistingPegInPegChoices + 1) % pegChoices.count]
             guess.pegs[index] = newPeg
         } else {
-            guess.pegs[index] = pegChoices.first ?? Code.missingPeg
-        }
-    }
-}
-
-struct Code {
-    var kind: Kind
-    var pegs: [Peg]
-    
-    static let missingPeg: Peg = "clear"
-    
-    enum Kind: Equatable {
-        case master
-        case guess
-        case attempt([Match])
-        case unknown
-    }
-    
-    var hasMissingPegs: Bool { pegs.contains { $0 == Code.missingPeg } }
-    
-    init(kind: Kind, numberOfPegs: Int) {
-        self.kind = kind
-        self.pegs = Array(repeating: Code.missingPeg, count: numberOfPegs)
-    }
-    
-    mutating func randomize(from pegChoices: [Peg]) {
-        for index in pegs.indices {
-            pegs[index] = pegChoices.randomElement() ?? Code.missingPeg
-        }
-    }
-    
-    var matches: [Match]? {
-        switch kind {
-        case .attempt(let matches): matches
-        default: nil
-        }
-    }
-    
-    func match(against otherCode: Code) -> [Match] {
-        var pegsToMatch = otherCode.pegs
-        
-        let backwardsExactMatches = pegs.indices.reversed().map { index in
-            if pegsToMatch.count > index, pegsToMatch[index] == pegs[index] {
-                pegsToMatch.remove(at: index)
-                return Match.exact
-            } else {
-                return .nomatch
-            }
-        }
-        
-        let exactMatches = Array(backwardsExactMatches.reversed())
-        return pegs.indices.map { index in
-            if exactMatches[index] != .exact, let matchIndex = pegsToMatch.firstIndex(of: pegs[index]) {
-                pegsToMatch.remove(at: matchIndex)
-                return .inexact
-            } else {
-                return exactMatches[index]
-            }
+            guess.pegs[index] = pegChoices.first ?? Peg.missing
         }
     }
 }
